% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_name.R
\name{new_name}
\alias{new_name}
\title{Title}
\usage{
new_name(old, new, board)
}
\arguments{
\item{old}{String. The name of the old object.}

\item{new}{String. The name of the new object.}

\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_rsconnect]{board_rsconnect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}
}
\value{
The object stored as 'new'.
}
\description{
Title
}
\examples{
board <- pins::board_temp()
reuse(1, "old", board)
new_name("old", "new", board)
reuse(name = "new")
}
