% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuse-package.R
\docType{package}
\name{reuse-package}
\alias{reuse}
\alias{reuse-package}
\title{reuse: Reuse R Objects}
\description{
Reuse R objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/reuse}
  \item \url{https://maurolepore.github.io/reuse/}
  \item Report bugs at \url{https://github.com/maurolepore/reuse/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

}
\keyword{internal}
