% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_name.R
\name{new_name}
\alias{new_name}
\title{Rename an R object you're reusing}
\usage{
new_name(old, new, board)
}
\arguments{
\item{old, new}{String. The name of the old and new objects.}

\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_rsconnect]{board_rsconnect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}
}
\value{
The object stored as 'new'.
}
\description{
Rename an R object you're reusing
}
\examples{
library(pins)

board <- board_temp()
x <- 1

x \%>\% reuse("old", board)
pin_exists("old", board = board)

"old" \%>\% new_name("new", board)
pin_exists("old", board = board)
pin_exists("new", board = board)
}
