% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuse.R
\name{reuse}
\alias{reuse}
\alias{board_reuse}
\title{Reuse R objects}
\usage{
reuse(
  x,
  name,
  board = board_reuse(),
  type = "qs",
  update = getOption("reuse.update", default = FALSE)
)

board_reuse()
}
\arguments{
\item{x}{An object (typically a data frame) to pin.}

\item{name}{Pin name.}

\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_rsconnect]{board_rsconnect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}

\item{type}{File type used to save \code{x} to disk. Must be one of
"csv", "rds", "json", "arrow", or "qs". If not supplied will use json for bare
lists and rds for everything else.}

\item{update}{Logical. update the cache?}
}
\value{
The R object written to or read from the cache.
}
\description{
Reuse R objects
}
\examples{
one <- 1 \%>\% reuse("one")
two <- 2 \%>\% reuse("one")
two

two <- 2 \%>\% reuse("one", update = TRUE)
two
withr::local_options(list(reuse.board = pins::board_temp()))
board_reuse()

pins::pin_exists(board_reuse(), name = "one")
1 \%>\% reuse("one")
pins::pin_exists(board_reuse(), name = "one")
}
