% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuse.R
\name{reuse}
\alias{reuse}
\title{Reuse R objects if possible}
\usage{
reuse(
  x,
  name,
  board = pins::board_folder(rappdirs::user_cache_dir("reuse")),
  overwrite = getOption("reuse.overwrite", default = FALSE)
)
}
\arguments{
\item{x}{An object (typically a data frame) to pin.}

\item{name}{Pin name.}

\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_rsconnect]{board_rsconnect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}

\item{overwrite}{Logical. Overwrite the cache?}
}
\value{
The R object written to or read from the cache.
}
\description{
Reuse R objects if possible
}
\examples{
one <- 1 \%>\% reuse("one")
two <- 2 \%>\% reuse("one")
two

two <- 2 \%>\% reuse("one", overwrite = TRUE)
two
}
